Finish task list editor
