+++ package.json
@@ -21,9 +21,12 @@
         "@internationalized/date": "^3.11.0",
         "@nuxt/ui": "^4.5.1",
         "@sidebase/nuxt-auth": "^1.2.0",
+        "@vueuse/core": "^14.2.1",
+        "@vueuse/integrations": "^14.2.1",
         "axios": "^1.13.5",
         "lucide": "^0.576.0",
         "simple-icons": "^16.10.0",
+        "sortablejs": "^1.15.7",
         "tailwindcss": "^4.1.18",
         "valibot": "^1.2.0",
         "vue": "^3.5.28",
